+++ sensor_msgs/test/test_pointcloud_iterator.cpp
@@ -45,7 +45,7 @@ TEST(sensor_msgs, PointCloud2Iterator)
   // Create a dummy PointCloud2
   size_t n_points = 4;
   sensor_msgs::msg::PointCloud2 cloud_msg_1, cloud_msg_2;
-  cloud_msg_1.height = n_points;
+  cloud_msg_1.height = static_cast<uint32_t>(n_points);
   cloud_msg_1.width = 1;
   sensor_msgs::PointCloud2Modifier modifier(cloud_msg_1);
   modifier.setPointCloud2FieldsByString(2, "xyz", "rgb");
